+++ src/main/java/com/trax/services/user/UserService.java
@@ -16,9 +16,9 @@ public interface UserService {
 
     public void addUser(User user);
     public void updateUser(User user);
-    public User getUser(int id);
+    public User getUser(Long id);
     public User getUser(String username);
-    public void deleteUser(int id);
+    public void deleteUser(Long id);
     public List getUsers();
 
 }
